configやrouterの設定の整理&controller配下にadminディレクトリ作成
